requirements.txt
base58==2.1.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cycler==0.12.1
fastapi==0.95.1
h5py==3.12.1
httpx==0.24.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
Pillow==11.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==1.10.7
Pympler==1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pyjwt[crypto]==2.6.0
PyJWT==2.6.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
setuptools==75.5.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.22.0

# liveness:
av==13.1.0
openai-whisper==20240930
geopy==2.4.1